Bulk changes for more explicit petagene usage
Bulk changes for more explicit petagene usage, we now leverage a runner_preset which should fusemount petagene at the per task singularity level.
We also are now removing fastqs after alignment has finished, this is a remnant from a separate workaround that was planned, but still is a good
feature to have for reducing our project footprint.